fx logic error in create auction
