+++ src/hooks/startfiNft.ts
@@ -3,7 +3,7 @@ import { useSubmitTransaction } from 'services/Blockchain/submitTransaction'
 import { useWalletModalToggle } from 'state/application/hooks'
 import { evaluateTransaction } from 'services/Blockchain/useEvaluateTransaction'
 import { useActiveWeb3React } from 'hooks'
-import { useStartFiPayment, useStartFiNft, parseBigNumber } from './useContract'
+import {  useStartFiNft, parseBigNumber } from './useContract'
 import { ROLES } from 'constants/index'
 import abiDecoder from 'abi-decoder'
 import { abi as _ABI } from '../constants/abis/StartfiRoyaltyNFT.json'
@@ -25,44 +25,44 @@ export const useNftInfo = () => {
   }, [contract])
 }
 
-export const useMint = (): ((
-  address: string,
-  ipfsHash: string,
-  share?: string | number,
-  base?: string | number
-) => any) => {
-  const { account, library } = useActiveWeb3React()
-  const contract = useStartFiPayment(true)
-  const mint = useSubmitTransaction()
-  const toggleWalletModal = useWalletModalToggle()
-  return useCallback(
-    async (address: string, ipfsHash: string, share?: string | number, base?: string | number) => {
-      if (!account || !address) {
-        toggleWalletModal()
-        return `account: ${account} is not connected`
-      }
-      try {
-        if (share && base) {
-          const mintedNFT = await mint(
-            'MintNFTWithRoyalty',
-            [address, ipfsHash, share, base],
-            contract,
-            account,
-            library
-          )
-          return (mintedNFT as any).value.toNumber()
-        } else {
-          const mintedNFT = await mint('MintNFTWithoutRoyalty', [address, ipfsHash], contract, account, library)
-          return (mintedNFT as any).value.toNumber()
-        }
-      } catch (e) {
-        console.log('error', e)
-        return e
-      }
-    },
-    [account, contract, library, mint, toggleWalletModal]
-  )
-}
+// export const useMint = (): ((
+//   address: string,
+//   ipfsHash: string,
+//   share?: string | number,
+//   base?: string | number
+// ) => any) => {
+//   const { account, library } = useActiveWeb3React()
+//   const contract = useStartFiPayment(true)
+//   const mint = useSubmitTransaction()
+//   const toggleWalletModal = useWalletModalToggle()
+//   return useCallback(
+//     async (address: string, ipfsHash: string, share?: string | number, base?: string | number) => {
+//       if (!account || !address) {
+//         toggleWalletModal()
+//         return `account: ${account} is not connected`
+//       }
+//       try {
+//         if (share && base) {
+//           const mintedNFT = await mint(
+//             'MintNFTWithRoyalty',
+//             [address, ipfsHash, share, base],
+//             contract,
+//             account,
+//             library
+//           )
+//           return (mintedNFT as any).value.toNumber()
+//         } else {
+//           const mintedNFT = await mint('MintNFTWithoutRoyalty', [address, ipfsHash], contract, account, library)
+//           return (mintedNFT as any).value.toNumber()
+//         }
+//       } catch (e) {
+//         console.log('error', e)
+//         return e
+//       }
+//     },
+//     [account, contract, library, mint, toggleWalletModal]
+//   )
+// }
 
 export const useGetTokenURI = (): ((tokenId: string | number) => any) => {
   const contract = useStartFiNft(false)
@@ -199,83 +199,83 @@ export const useRoyaltyInfo = (): ((tokenId: string | number, value: string | nu
   )
 }
 
-export const useNftPaymentInfo = (): (() => any) => {
-  const contract = useStartFiPayment(false)
-  return useCallback(() => {
-    const getInfo = async () => {
-      try {
-        const info = await evaluateTransaction(contract, 'info', [])
-        return parseBigNumber(info)
-      } catch (e) {
-        console.log(e)
-      }
-    }
-    return getInfo()
-  }, [contract])
-}
+// export const useNftPaymentInfo = (): (() => any) => {
+//   const contract = useStartFiPayment(false)
+//   return useCallback(() => {
+//     const getInfo = async () => {
+//       try {
+//         const info = await evaluateTransaction(contract, 'info', [])
+//         return parseBigNumber(info)
+//       } catch (e) {
+//         console.log(e)
+//       }
+//     }
+//     return getInfo()
+//   }, [contract])
+// }
 // NFT Payment Owner Transactions
-export const useChangeFeesNftPayment = (): ((newFees: string | number) => any) => {
-  const { account, library } = useActiveWeb3React()
-  const contract = useStartFiPayment(true)
-  const changeFees = useSubmitTransaction()
-  const toggleWalletModal = useWalletModalToggle()
-  return useCallback(
-    async (newFees: string | number) => {
-      if (!account) {
-        toggleWalletModal()
-        return `account: ${account} is not connected`
-      }
-      try {
-        return await changeFees('changeFees', [newFees], contract, account, library)
-      } catch (e) {
-        console.log('error', e)
-        return e
-      }
-    },
-    [account, contract, library, changeFees, toggleWalletModal]
-  )
-}
+// export const useChangeFeesNftPayment = (): ((newFees: string | number) => any) => {
+//   const { account, library } = useActiveWeb3React()
+//   const contract = useStartFiPayment(true)
+//   const changeFees = useSubmitTransaction()
+//   const toggleWalletModal = useWalletModalToggle()
+//   return useCallback(
+//     async (newFees: string | number) => {
+//       if (!account) {
+//         toggleWalletModal()
+//         return `account: ${account} is not connected`
+//       }
+//       try {
+//         return await changeFees('changeFees', [newFees], contract, account, library)
+//       } catch (e) {
+//         console.log('error', e)
+//         return e
+//       }
+//     },
+//     [account, contract, library, changeFees, toggleWalletModal]
+//   )
+// }
 
-export const useChangeNftContractNftPayment = (): ((nftAddress: string) => any) => {
-  const { account, library } = useActiveWeb3React()
-  const contract = useStartFiPayment(true)
-  const changeNftContract = useSubmitTransaction()
-  const toggleWalletModal = useWalletModalToggle()
-  return useCallback(
-    async (nftAddress: string) => {
-      if (!account) {
-        toggleWalletModal()
-        return `account: ${account} is not connected`
-      }
-      try {
-        return await changeNftContract('changeNftContract', [nftAddress], contract, account, library)
-      } catch (e) {
-        console.log('error', e)
-        return e
-      }
-    },
-    [account, contract, library, changeNftContract, toggleWalletModal]
-  )
-}
+// export const useChangeNftContractNftPayment = (): ((nftAddress: string) => any) => {
+//   const { account, library } = useActiveWeb3React()
+//   const contract = useStartFiPayment(true)
+//   const changeNftContract = useSubmitTransaction()
+//   const toggleWalletModal = useWalletModalToggle()
+//   return useCallback(
+//     async (nftAddress: string) => {
+//       if (!account) {
+//         toggleWalletModal()
+//         return `account: ${account} is not connected`
+//       }
+//       try {
+//         return await changeNftContract('changeNftContract', [nftAddress], contract, account, library)
+//       } catch (e) {
+//         console.log('error', e)
+//         return e
+//       }
+//     },
+//     [account, contract, library, changeNftContract, toggleWalletModal]
+//   )
+// }
 
-export const useChangePaymentContractNftPayment = (): ((paymentAddress: string) => any) => {
-  const { account, library } = useActiveWeb3React()
-  const contract = useStartFiPayment(true)
-  const changePaymentContract = useSubmitTransaction()
-  const toggleWalletModal = useWalletModalToggle()
-  return useCallback(
-    async (nftAddress: string) => {
-      if (!account) {
-        toggleWalletModal()
-        return `account: ${account} is not connected`
-      }
-      try {
-        return await changePaymentContract('changePaymentContract', [nftAddress], contract, account, library)
-      } catch (e) {
-        console.log('error', e)
-        return e
-      }
-    },
-    [account, contract, library, changePaymentContract, toggleWalletModal]
-  )
-}
+// export const useChangePaymentContractNftPayment = (): ((paymentAddress: string) => any) => {
+//   const { account, library } = useActiveWeb3React()
+//   const contract = useStartFiPayment(true)
+//   const changePaymentContract = useSubmitTransaction()
+//   const toggleWalletModal = useWalletModalToggle()
+//   return useCallback(
+//     async (nftAddress: string) => {
+//       if (!account) {
+//         toggleWalletModal()
+//         return `account: ${account} is not connected`
+//       }
+//       try {
+//         return await changePaymentContract('changePaymentContract', [nftAddress], contract, account, library)
+//       } catch (e) {
+//         console.log('error', e)
+//         return e
+//       }
+//     },
+//     [account, contract, library, changePaymentContract, toggleWalletModal]
+//   )
+// }
